+++ internal/av1/loopfilter/filter4_test.go
@@ -128,9 +128,9 @@ func TestFilter4EdgeRejectsInvalidInputs(t *testing.T) {
 		bytesPerSample int
 		bitDepth       uint8
 		edge           Edge
-		x              int
-		y              int
-		length         int
+		x              int32
+		y              int32
+		length         int32
 	}{
 		{name: "sample width", plane: plane, bytesPerSample: 2, bitDepth: 8, edge: EdgeHorizontal, x: 0, y: 2, length: 4},
 		{name: "edge", plane: plane, bytesPerSample: 1, bitDepth: 8, edge: Edge(9), x: 0, y: 2, length: 4},
@@ -198,7 +198,7 @@ func FuzzFilter4Edge(f *testing.F) {
 		if edge == EdgeVertical {
 			length = height - y
 		}
-		err := Filter4Edge(plane, bytesPerSample, bitDepth, edge, x, y, length, Thresholds{
+		err := Filter4Edge(plane, bytesPerSample, bitDepth, edge, int32(x), int32(y), int32(length), Thresholds{
 			Limit:            rawLimit,
 			BlockLimit:       rawBlockLimit,
 			HighEdgeVariance: rawHEV,
