+++ internal/av1/loopfilter/filter6_test.go
@@ -187,7 +187,7 @@ func TestFilter6EdgeMatchesCReference(t *testing.T) {
 			writeFilter6ReferenceSample(want, bytesPerSample, tt.edge, i, ref)
 		}
 
-		x, y, length := 0, 3, len(tt.samples)
+		x, y, length := int32(0), int32(3), int32(len(tt.samples))
 		if tt.edge == EdgeVertical {
 			x, y = 3, 0
 		}
@@ -208,9 +208,9 @@ func TestFilter6EdgeRejectsInvalidInputs(t *testing.T) {
 		bytesPerSample int
 		bitDepth       uint8
 		edge           Edge
-		x              int
-		y              int
-		length         int
+		x              int32
+		y              int32
+		length         int32
 	}{
 		{name: "sample width", plane: plane, bytesPerSample: 2, bitDepth: 8, edge: EdgeHorizontal, x: 0, y: 3, length: 4},
 		{name: "edge", plane: plane, bytesPerSample: 1, bitDepth: 8, edge: Edge(9), x: 0, y: 3, length: 4},
@@ -278,7 +278,7 @@ func FuzzFilter6Edge(f *testing.F) {
 		if edge == EdgeVertical {
 			length = height - y
 		}
-		err := Filter6Edge(plane, bytesPerSample, bitDepth, edge, x, y, length, Thresholds{
+		err := Filter6Edge(plane, bytesPerSample, bitDepth, edge, int32(x), int32(y), int32(length), Thresholds{
 			Limit:            rawLimit,
 			BlockLimit:       rawBlockLimit,
 			HighEdgeVariance: rawHEV,
